Policy (docs/reconcile/2026-08-26-merge-resolution.md): main is the prior-revision oracle for
every content conflict and for the 9 add/add modules with no branch-only consumers of HEAD-only
symbols; HEAD's version of each is preserved under docs/reconcile/vss-head-versions/. Three
add/add modules keep HEAD (branch code imports ActionPolicyResult / ZKProof / TaskClass) with
main's VerificationResult / ZKFVProof / RoutingResult+TIER_* appended (ast-extracted) so main-side
importers (agi/__init__, contracts, swarm/tip_of_spear_router) keep working — import-smoked.
Markdown/config unioned; scripts and uv.lock take main (ruff campaign); branch hunks preserved.
Triage evidence: vault reports/20260826-vss-conflict-triage-qwen3.{jsonl,md} (Qwen3.6-35B judge,
Gemma-4-E4B refute lane, 4/4 sampled claims grep-verified), depcheck txt.

+++ src/cohezion/arc/transforms.py
@@ -972,8 +972,8 @@ def object_center_of_mass(grid: np.ndarray) -> np.ndarray | None:
     for color in colors:
         c_coords = np.argwhere(grid == color)
         if len(c_coords) > 0:
-            mean_r = int(round(float(np.mean(c_coords[:, 0]))))
-            mean_c = int(round(float(np.mean(c_coords[:, 1]))))
+            mean_r = round(float(np.mean(c_coords[:, 0])))
+            mean_c = round(float(np.mean(c_coords[:, 1])))
             mean_r = min(max(0, mean_r), grid.shape[0] - 1)
             mean_c = min(max(0, mean_c), grid.shape[1] - 1)
             result[mean_r, mean_c] = color
